Github actions and docker
Now the .env will be created also on the VM on github when doing the push using the github secrets. It also creates the keystore.p12 to keep the system secure.

The docker-compose was updated to include again the version. It gives a warning, but without it docker simply can't run on github.