Completed artifact creation
Finally brought to a proper working state the creation of an artifact. Artifacts will be stored in the directory "artifacts/" located in the root of the project. Here every file will start with a unique name. The file metadata instead will be stored in MongoDB in the apposite collection.

Also did some small fixes on GmailService and to the setup scripts, to put some space here and there and to handle correctly the presence of an already existing setup.zip folder when copying the secrets from the project.